% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{corr_exp}
\alias{corr_exp}
\title{Exponential correlation matrix}
\usage{
corr_exp(distance, lambda, jitter = FALSE, jitter_value = 0.001)
}
\arguments{
\item{distance}{An JxJ matrix of distances between the J stations.}

\item{lambda}{length-scale or frequency parameter.}

\item{jitter}{add a jitter to the correlation matrix diagonal if jitter = TRUE.}

\item{jitter_value}{jitter value.}
}
\value{
a real JxJ correlation matrix.
}
\description{
Produce an exponential correlation matrix based on a distance matrix.
}
\examples{
#Generate a random set of 100 stations coordinates for the example's purpose
coord = data.frame(latitude = runif(100,-25,25), longitude = runif(100, -25, 25))
#Compute the exponential correlation matrix with fixed lambda and added jitter
corr_exp(coord, 0.5, jitter = FALSE, 0.001)
}
