% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{CRPS_approx}
\alias{CRPS_approx}
\title{Approximate CRPS}
\usage{
CRPS_approx(y_sample, y_obs)
}
\arguments{
\item{y_sample}{sample of predictions for the observation.}

\item{y_obs}{observation value.}
}
\value{
numeric CRPS value.
}
\description{
Compute the approximate CRPS from a sample of prediction based on the generalized quantile function.
}
