% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOS_models.R
\name{CRPS_emos}
\alias{CRPS_emos}
\title{Mean CRPS function of the EMOS model.}
\usage{
CRPS_emos(par, x, s, y)
}
\arguments{
\item{par}{Optimized parameters obtained on training data with BFGS method, for example.}

\item{x}{testing data NWP means.}

\item{s}{testing data NWP variances.}

\item{y}{testing data observed values.}
}
\value{
mean CRPS value.
}
\description{
Computes the mean CRPS for the EMOS model based on numerical weather predictions (NWP), using the exact CRPS formula for the normal distribution.
}
