% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOS_models.R
\name{CRPS_emos_list}
\alias{CRPS_emos_list}
\title{CRPS function of the EMOS model.}
\usage{
CRPS_emos_list(par, x, s, y)
}
\arguments{
\item{par}{Optimized parameters obtained on training data with BFGS method, for example.}

\item{x}{testing data NWP means.}

\item{y}{testing data observed values.}
}
\value{
List with vector of CRPS and mean CRPS value.
}
\description{
Computes each observation CRPS with the mean CRPS for the EMOS model based on numerical weather predictions (NWP),
 using the exact CRPS formula for the normal distribution.
}
