% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{Empirical_Variance}
\alias{Empirical_Variance}
\title{Empirical Variance}
\usage{
Empirical_Variance(X)
}
\arguments{
\item{X}{data.frame or matrix such that rows=observations and columns=variables.}
}
\value{
covariance matrix.
}
\description{
Compute the empirical variance from vectors of observations.
}
