% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_reg_models.R
\name{global_Bayes_reg}
\alias{global_Bayes_reg}
\title{Global Bayesian regression model}
\usage{
global_Bayes_reg(
  y_test,
  X_test,
  X_train,
  y_train,
  beta_0,
  V_0,
  a_0,
  b_0,
  approx = FALSE,
  nb_draws = 200,
  plot_hist_PIT = FALSE
)
}
\arguments{
\item{y_test}{vector of observations responses from the testing data.}

\item{X_test}{matrix of observations covariates from the testing data used for regression (rows = observations, columns = covariates, first column = unit vector).}

\item{X_train}{matrix of observations covariates from the training data and used for regression (rows = observations, columns = covariates, first column = unit vector).}

\item{y_train}{vector of observations responses from the training data.}

\item{beta_0}{mean of the Gaussian prior.}

\item{V_0}{covariance matrix of the Gaussian prior.}

\item{a_0}{shape parameter of the inverse-gamma prior.}

\item{b_0}{rate parameter of the inverse-gamma prior.}

\item{approx}{if approx is TRUE, approximate the CRPS instead of using the exact formula to compute it. Recommended to always use approx = FALSE, since
approximate method is for examples only.}

\item{nb_draws}{if approx is TRUE, number of draw from the estimated predictive distribution used to approximate the CRPS.}

\item{plot_hist_PIT}{if plot_hist_PIT is TRUE, plot the PIT histogram.}
}
\value{
data.frame with the pair of observed/predicted values with associated AE, CRPS and PIT metrics.
}
\description{
Run the global Bayesian regression model and get predicted values with mean average error and exact CRPS.
}
