% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOS_models.R
\name{global_mean_EMOS}
\alias{global_mean_EMOS}
\title{Global mean-EMOS model.}
\usage{
global_mean_EMOS(df_train, df_test)
}
\arguments{
\item{df_train}{training data.frame/matrix with following columns: space and time indexes as columns "time" and "station", "x" and "s" columns as NWP means and variances
and "y" as observed training values.}

\item{df_test}{testing dataframe/matrix with the same columns as training data.}
}
\value{
list containing testing observations/predictions pairs with their associated CRPS/AE and PIT metrics and the mean CRPS and MAE of all testing data.
}
\description{
Run the global mean-EMOS model with the BFGS method.
}
