% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STPPMX_models.R
\name{grid_matrices}
\alias{grid_matrices}
\title{Compute relevant matrices on a parameter grid}
\usage{
grid_matrices(
  grid_varphi,
  lambda,
  W,
  grid,
  jitter_H = FALSE,
  jitter_varphi = TRUE
)
}
\arguments{
\item{grid_varphi}{vector of time correlation parameter values such that we calculates the matrices for each.}

\item{lambda}{space correlation parameter.}

\item{W}{Distance matrix between stations coordinates.}

\item{grid}{data.frame or matrix with rows = observations, first column = time points and second columns = stations.}

\item{jitter_H}{if TRUE, add a jitter to the diagonal of the space correlation matrix.}

\item{jitter_varphi}{if TRUE, add a jitter to the diagonal of the time correlation matrix.}
}
\value{
List containing, for each varphi, matrices with 2 columns each containing matrices calculated for each time point.
}
\description{
Pre-compute two relevant matrices to be then used in the STPPMx model with mixed effects. Like explained in my paper, this allows to greatly speed up
mixed effects STPPMx model by pre-calculating fastidious matrix for a fixed grid of selected parameters.
}
