% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOS_models.R
\name{hist_PIT}
\alias{hist_PIT}
\title{PIT histogram}
\usage{
hist_PIT(PIT, nb_breaks)
}
\arguments{
\item{PIT}{vector of PIT values.}

\item{nb_breaks}{number of histogram breaks desired.}
}
\value{
PIT histogram plot.
}
\description{
Creates a PIT histogram based on given PIT values and numbers of breaks wanted.
}
