% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Importance}
\alias{Importance}
\title{Importance}
\usage{
Importance(CRPS_FULL, CRPS_shuffled)
}
\arguments{
\item{CRPS_FULL}{CRPS computed from the full model.}

\item{CRPS_shuffled}{CRPS computed from the shuffled model.}
}
\value{
Importance numeric value.
}
\description{
Compute covariates importance based on CRPS.
}
