% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{L2}
\alias{L2}
\title{L2 distance}
\usage{
L2(x, y)
}
\arguments{
\item{x}{a vector.}

\item{y}{a vector.}
}
\value{
a real number.
}
\description{
Compute the L2 or euclidean distance between two numeric vectors.
}
\examples{
L2(c(-1,-1), c(1,2))
}
