% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOS_models.R
\name{local_mean_EMOS}
\alias{local_mean_EMOS}
\title{Local mean-EMOS model.}
\usage{
local_mean_EMOS(df_train, df_test)
}
\arguments{
\item{df_train}{training dataframe/matrix with following columns: spacetime indexes as columns "time" and "station", "x" and "s" columns as NWP means and variances
and "y" as observed training values.}

\item{df_test}{testing dataframe/matrix with the same columns as training data.}
}
\value{
list containing testing observations/predictions pairs with their associated CRPS/AE and PIT metrics, space points mean CRPS/MAE over time points
and the mean CRPS and MAE of all testing data.
}
\description{
Run the Local mean-EMOS model with the BFGS method. Here, local refers to having an individual models for each stations, meaning that all time points
associated with a station share the same model, opposed to the global model where all observations share the same model regardless of spacetime.
}
