% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOS_models.R
\name{MAE_emos}
\alias{MAE_emos}
\title{MAE function of the EMOS model.}
\usage{
MAE_emos(par, x, y)
}
\arguments{
\item{par}{Optimized parameters obtained on training data with BFGS method, for example.}

\item{x}{testing data NWP means.}

\item{y}{testing data observed values.}
}
\value{
MAE value.
}
\description{
Computes the mean absolute error (MAE) for the EMOS model based on numerical weather predictions (NWP).
}
