% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOS_models.R
\name{MAE_emos_list}
\alias{MAE_emos_list}
\title{AE function of the EMOS model.}
\usage{
MAE_emos_list(par, x, y)
}
\arguments{
\item{par}{Optimized parameters obtained on training data with BFGS method, for example.}

\item{x}{testing data NWP means.}

\item{y}{testing data observed values.}
}
\value{
list with vector of AE and MAE value.
}
\description{
Computes each observation average error (AE) with the mean absolute error (MAE) for the EMOS model based on numerical weather predictions (NWP).
}
