% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{MSE}
\alias{MSE}
\title{Mean squared error (MSE)}
\usage{
MSE(y_estimated, y_obs, absolute = FALSE)
}
\arguments{
\item{y_estimated}{estimated value for y_obs.}

\item{y_obs}{observation value.}

\item{absolute}{compute the mean absolute error instead (TRUE = mean absolute error, FALSE = mean squared error).}
}
\value{
numeric average value
}
\description{
Compute the mean squared error (MSE) or the mean absolute error (MAE).
}
