% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{multivariate_t_density}
\alias{multivariate_t_density}
\title{Multivariate student's t density}
\usage{
multivariate_t_density(x, v, mu, Sig, log_d = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{v}{degrees of freedom ( v > 0).}

\item{mu}{mean parameter vector.}

\item{Sig}{scale parameter matrix.}

\item{log_d}{compute the density on the log-scale if log_d is TRUE.}
}
\value{
numeric density value.
}
\description{
Evaluate the multivariate student's density at one point given the distribution parameters.
}
