% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{overlap_index}
\alias{overlap_index}
\title{Clusters overlap indexes}
\usage{
overlap_index(X, c)
}
\arguments{
\item{X}{matrix or data.frame with columns = variables and rows=observations.}

\item{c}{vector of clusters indicator values.}
}
\value{
a data.frame containing the metrics.
}
\description{
Function that computes the homogeneity, the separation and the similarity index in clustered data.
}
