% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{Simulation_MixSim}
\alias{Simulation_MixSim}
\title{Simulate clustered data using MixSim.}
\usage{
Simulation_MixSim(
  max_overlap,
  bar_overlap,
  lim_x,
  K,
  N,
  nb_clusters,
  p,
  tau_time,
  varphi,
  tau_space,
  station_by_time = TRUE,
  distribution,
  sigma_2,
  add_space = FALSE,
  add_time = TRUE
)
}
\arguments{
\item{max_overlap}{maximum overlap used for the controlling overlap.}

\item{bar_overlap}{average overlap used for the controlling overlap.}

\item{lim_x}{specific boundaries on the space of each covariate dimension. By default (0,1).}

\item{K}{number of stations/space points.}

\item{N}{number of time points.}

\item{nb_clusters}{number of clusters.}

\item{p}{covariates dimension size.}

\item{tau_time}{variance parameter for the random time effects.}

\item{varphi}{correlation parameter for the random time effects.}

\item{tau_space}{variance parameter for the spatial random effects.}

\item{station_by_time}{how to structure the long data grid. If TRUE, list each station for time 1, then each station for time 2, etc...
If FALSE, do the opposite.}

\item{distribution}{"N" for normally distributed response, "LN" for a log-normal distribution instead.}

\item{sigma_2}{response variance parameter. Used to control noise in the response value.}

\item{add_space}{if TRUE add spatial random effects.}

\item{add_time}{if TRUE, add time random effects.}
}
\value{
a list containing simulated training and test data, generated coordinates, distance matrix W and parameters used to generate the observations.
}
\description{
Simulate clustered data using MixSim by specifying the average and/or max overlap parameters. Given the chosen number of clusters,
stations, time points and dimensions for the covariates, simulate clustered data with covariate and response value, with chosen added spatiotemporal random effects.
Also generates random coordinates on an uniform square.
}
