% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{Simulation_real_data}
\alias{Simulation_real_data}
\title{Simulate clustered data based on an existing dataset.}
\usage{
Simulation_real_data(
  clean_data,
  idx_train,
  idx_test,
  station_info,
  nb_clusters,
  p,
  K,
  idx_time,
  distribution,
  same_stations = TRUE,
  add_time = TRUE,
  time_sorted = FALSE,
  mapping.daymonth,
  disp_beta,
  nb_rep,
  sigma_time,
  factor_sigma,
  plot_ALL = FALSE,
  jitter_coord = FALSE,
  jitter_cov = 0.001,
  plot.coord = FALSE
)
}
\arguments{
\item{clean_data}{Clean dataset from which you want to generate clustered simulated datasets. Needs to have the following columns: "date", "station",
"time", "obs" and "t2m_mean".
Rest of the columns are other covariates.}

\item{idx_train}{vector of training data rows indexes.}

\item{idx_test}{vector of testing data rows indexes.}

\item{station_info}{data.frame containing stations id's, latitudes and longitudes as the following columns respectively: "stations_id", "lat" and "lon".}

\item{nb_clusters}{number of clusters in simulated dataset.}

\item{p}{number of total covariates used. Since we use "t2m_mean", "lat", "lon", this determines hot many p-3 covariates are sampled based on correlation with obs.}

\item{K}{number of stations wanted in the simulated dataset (sampled uniformly from all stations).}

\item{idx_time}{vector of selected time wanted in the simulated dataset.}

\item{distribution}{simulated response distribution ("N" for normal distribution or "LN" for log-normal)}

\item{same_stations}{if TRUE, keep times with a lot of stations and then keep stations present accross all selected times (used to make sure each station has
the same time points)}

\item{add_time}{if TRUE, add time effects (randomly distributed around a sin function).}

\item{time_sorted}{if TRUE, rearrange the N time points to a 1:N sequence, otherwise keep the original indexing.}

\item{mapping.daymonth}{mapping matrix to generate the random time effects.}

\item{disp_beta}{dispersion parameter for the random walk generation of the beta parameters.}

\item{nb_rep}{number of steps in the random walk.}

\item{sigma_time}{variance parameters for the simulated time effects.}

\item{factor_sigma}{variance parameter for the randomly generated betas parameters.}

\item{plot_ALL}{if TRUE, plot generated random time effects and noises.}

\item{jitter_coord}{if TRUE, add a jitter to the coordinates.}

\item{jitter_cov}{if TRUE, add a jitter to the simulated covariates covariance matrix.}

\item{plot.coord}{if TRUE, plot the selected stations coordinates.}

\item{idx_covariate}{columns indexing which covariates in clean_data to be sampled for simulation, other than "t2m_mean", "lat" and "lon" which are
automatically included.}
}
\value{
a list containing the simulated dataset, the scaled coordinates, the clusters indicators vectors and the generated time effects vector.
}
\description{
Simulate clustered data based on an existing dataset.
}
