% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{sin_embedding}
\alias{sin_embedding}
\title{Sinusoidal embedding}
\usage{
sin_embedding(x, period)
}
\arguments{
\item{x}{a vector of time points.}

\item{period}{period of the sin function.}
}
\value{
a numeric vector with the sin embedding.
}
\description{
Create a sinusoidal embedding from a vector of time points and a period parameter.
}
