\name{STPPMx-package}
\alias{STPPMx-package}
\alias{STPPMx}
\docType{package}
\title{
The STPPMx package.
}
\description{
Provides functions to run STPPMx and associated post-processing models described in my master's thesis (Link below).
}
\details{
Many functions are coded in C++ with the help of Rcpp and RcppArmadillo. All the code in  the package were used in the context of my master's thesis academic reasearch published on Papyrus (link and citation in package). If you start reading the code, try to read the thesis at the same time to understand were my function are coming from. Also, if you use the code in any of your work, I would appreciate if you could cite my paper and mention that you used my code!
}
\author{
Rayane Dounas
}
\references{
Dounas, R. (2025). Prévisions météorologiques par régression non-paramétrique bayésienne basée sur le modèle PPMx spatio-temporel. [Master's thesis, Université de Montréal]. Papyrus. https://hdl.handle.net/1866/41129.

Conrad Sanderson and Ryan Curtin. Armadillo: An Efficient Framework for Numerical Linear Algebra. International Conference on Computer and Automation Engineering, 2025.
https://doi.org/10.1109/ICCAE64891.2025.10980539.
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = anRpackage)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{anRpackage}}
%\section{Documentation Index}{\packageIndices{anRpackage}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
