% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STPPMX_models.R
\name{STPPMx}
\alias{STPPMx}
\title{STPPMX model}
\usage{
STPPMx(
  W,
  y,
  X_sim,
  X_reg,
  grid,
  nb_clusters_start,
  alpha,
  v,
  s,
  m_0,
  Phi,
  w,
  M,
  B,
  R,
  tau_time,
  tau_station,
  tau,
  lambda,
  varphi,
  MC_param,
  MH_vec,
  jitter_lambda = TRUE,
  jitter_varphi = TRUE,
  add_time = FALSE,
  add_station = FALSE,
  add_mixed = FALSE,
  precalc = TRUE,
  fast = FALSE,
  print.state = TRUE
)
}
\arguments{
\item{W}{distance matrix between stations coordinates.}

\item{y}{vector of training response values.}

\item{X_sim}{training covariates matrix used for similarity (rows = observations, columns = covariates).}

\item{X_reg}{training covariates matrix used for regression (rows = observations, columns = covariates, first column = unit vector).}

\item{grid}{data.frame or matrix with rows = observations, first column = time points and second columns = stations.}

\item{nb_clusters_start}{number of starting clusters.}

\item{alpha}{fixed alpha parameter for the Dirichlet process (Could be used as a starting parameter in the future).}

\item{v}{shape parameter of the inverse-gamma prior.}

\item{s}{rate parameter of the inverse-gamma prior.}

\item{m_0}{mean parameter for the similarity Gaussian prior.}

\item{Phi}{variance matrix parameter for the similarity inverse-Wishart prior.}

\item{w}{degrees of freedom for the similarity inverse-Wishart prior.}

\item{M}{number of auxiliary parameters to draw in Neal algorithm.}

\item{B}{number of MCMC iterations.}

\item{R}{number of burn-in iterations.}

\item{tau_time}{variance parameter for the inverse-gamma prior of the random time effects.}

\item{tau_station}{variance parameter for the inverse-gamma prior of the random space effects.}

\item{tau}{variance parameter for the inverse-gamma prior of the random spatio-temporal effects.}

\item{lambda}{space correlation parameter.}

\item{varphi}{time correlation parameter.}

\item{MC_param}{vector of TRUE/FALSE values indicating whether to use fixed value or MCMC draw for the
following parameters (in order) : varphi, lambda, tau_time, tau_station, alpha.}

\item{MH_vec}{vector containing values of the instrumental distribution step used for each of the above parameters (same order).}

\item{jitter_lambda}{if TRUE, add a jitter to the diagonal of the space correlation matrix.}

\item{jitter_varphi}{if TRUE, add a jitter to the diagonal of the time correlation matrix.}

\item{add_time}{if TRUE, use the random time effects model.}

\item{add_station}{if TRUE, use the random space effects model.}

\item{add_mixed}{if TRUE, use the random spatio-temporal effects model.}

\item{precalc}{if TRUE, use a grid of varphi values instead of a fixed value and pre-compute the associate matrices for speed.}

\item{fast}{if TRUE, use a faster Neal's algorithm partition sampling implementation that uses more memory to prevent redundant computations.}

\item{print.state}{if TRUE, print important information on the MCMC draws at each iteration.}
}
\value{
List containing MCMC draws and acceptance probabilities.
}
\description{
Run the STPPMX model MCMC training described in chapter 4 of my paper (citation in package description) and returns the MCMC parameters samples to be used with the
STPPMx_predict function to get useful metric from the predictive distribution.
}
