% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STPPMX_models.R
\name{STPPMx_predict}
\alias{STPPMx_predict}
\title{STPPMx model predictions and scores.}
\usage{
STPPMx_predict(
  X_train_sim,
  X_test_sim,
  X_test_reg,
  y_test,
  grid_test,
  MCMC_results,
  w,
  Phi,
  v,
  s,
  m_0,
  nb_draws
)
}
\arguments{
\item{X_train_sim}{training covariates matrix used for similarity (rows = observations, columns = covariates).}

\item{X_test_sim}{testing covariates matrix used for similarity (rows = observations, columns = covariates).}

\item{X_test_reg}{testing covariates matrix used for regression (rows = observations, columns = covariates, first column = unit vector).}

\item{y_test}{vector of testing response values.}

\item{grid_test}{data.frame or matrix with rows = testing observations, first column = time points and second columns = stations.}

\item{MCMC_results}{results list of MCMC parameters samples returned by the STPPMx function.}

\item{w}{degrees of liberty for the similarity inverse-Wishart prior.}

\item{Phi}{variance matrix parameter used in training for the similarity inverse-Wishart prior.}

\item{v}{shape parameter used in training of the inverse-gamma prior.}

\item{s}{rate parameter used in training of the inverse-gamma prior.}

\item{m_0}{mean parameter used in training for the similarity Gaussian prior.}

\item{nb_draws}{number of draws to approximate the predictive distribution using the method described in my thesis.}
}
\value{
list containing a data.frame with predicted values, CRPS and AE at each spacetime points, a vector with the mean CRPS and MAE and a matrix containing the
samples that were drawn to approximate the CRPS.
}
\description{
Get an STPPMx model predictions and scores (MAE/CRPS) based on results given by the STPPMx functions and testing data.
}
