% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{thresh_ind}
\alias{thresh_ind}
\title{Threshold indicator function}
\usage{
thresh_ind(x, y)
}
\arguments{
\item{x}{a numeric value.}

\item{y}{a numeric value.}
}
\value{
1 if x <= y, 0 if not.
}
\description{
Indicator function that gives 1 if x <= y, 0 if not.
}
